feat - use localStorage in thunks if not logged in
