+++ src/redux/actions/habit-local-storage.svc.ts
@@ -1,2 +1,20 @@
+import {ToggleLinkAction, CreateHabitAction, RemoveHabitAction, LoadDatesAction} from './habit-actions';
+
 const LOCAL_STORAGE_PREFIX = 'MOMENTUM_HABIT_DATA_';
 
+
+export function loadLocalHabits() {
+
+}
+
+export function createLocalHabit() {
+
+}
+
+export function toggleLocalHabit() {
+
+}
+
+export function removeLocalHabit() {
+
+}
